+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### [2.19.2] - 2022-04-12
+- Add payment detail note
+- fix note duplication on payment flow
+
 ### [2.19.1] - 2021-11-22
 - Change ecuador endpoint
 - Fix bug on payment result